remove old derive public keys code
