+++ RadixWallet/Clients/LedgerHardwareWalletClient/LedgerHardwareWalletClient+Interface.swift
@@ -5,7 +5,6 @@ struct LedgerHardwareWalletClient: Sendable {
 	var isConnectedToAnyConnectorExtension: IsConnectedToAnyConnectorExtension
 	var getDeviceInfo: GetDeviceInfo
 	var derivePublicKeys: DerivePublicKeys
-	var newDerivePublicKeys: NewDerivePublicKeys
 	var signTransaction: SignTransaction
 	var newSignTransaction: NewSignTransaction
 	var signPreAuthorization: SignPreAuthorization
@@ -18,8 +17,7 @@ struct LedgerHardwareWalletClient: Sendable {
 extension LedgerHardwareWalletClient {
 	typealias IsConnectedToAnyConnectorExtension = @Sendable () async -> AnyAsyncSequence<Bool>
 	typealias GetDeviceInfo = @Sendable () async throws -> P2P.ConnectorExtension.Response.LedgerHardwareWallet.Success.GetDeviceInfo
-	typealias DerivePublicKeys = @Sendable ([P2P.LedgerHardwareWallet.KeyParameters], LedgerHardwareWalletFactorSource) async throws -> [HierarchicalDeterministicPublicKey]
-	typealias NewDerivePublicKeys = @Sendable (DerivePublicKeysRequest) async throws -> [HierarchicalDeterministicFactorInstance]
+	typealias DerivePublicKeys = @Sendable (DerivePublicKeysRequest) async throws -> [HierarchicalDeterministicFactorInstance]
 
 	typealias DeriveAndDisplayAddress = @Sendable (P2P.LedgerHardwareWallet.KeyParameters, LedgerHardwareWalletFactorSource) async throws -> (HierarchicalDeterministicPublicKey, String)
