--- RadixWallet/Features/DerivePublicKeys/NewDerivePublicKeys+View.swift
@@ -1,24 +1,0 @@
-import SwiftUI
-
-// MARK: - NewDerivePublicKeys.View
-extension NewDerivePublicKeys {
-	struct View: SwiftUI.View {
-		let store: StoreOf<NewDerivePublicKeys>
-
-		var body: some SwiftUI.View {
-			WithPerceptionTracking {
-				if let child = store.scope(state: \.factorSourceAccess, action: \.child.factorSourceAccess) {
-					NewFactorSourceAccess.View(store: child)
-				} else {
-					// We don't need the FactorSourceAccess, so we can start the derivation right away
-					Rectangle()
-						.presentationDetents([.height(1)])
-						.presentationDragIndicator(.hidden)
-						.onFirstTask { @MainActor in
-							store.send(.internal(.deriveWithSpecificPrivateHD__MustImplement))
-						}
-				}
-			}
-		}
-	}
-}
